refactor(actions): split streaming reducers
- move MCP gate, lifecycle, error, tool-call, and sampling logic
  into focused action submodules
- keep `streaming.rs` as the streaming action dispatcher and route
  handlers through the new modules
- update tests to target module handlers directly for clearer
  ownership boundaries
- add action-handler architecture guidance to ARCHITECTURE and
  refresh README tree entries for recent refactors
- clarify AGENTS doc guidance: README for user-facing/tree updates,
  ARCHITECTURE for internal design notes as needed
