+++ src/core/app/actions/stream_errors.rs
@@ -1,17 +1,112 @@
 use super::{App, AppActionContext, AppCommand};
+use crate::core::message::AppMessageKind;
 
 pub(super) fn handle_stream_error(
     app: &mut App,
     message: String,
     ctx: AppActionContext,
 ) -> Option<AppCommand> {
-    super::handle_stream_error(app, message, ctx)
+    let error_message = message.trim().to_string();
+    if app.session.mcp_tools_enabled
+        && !app.session.mcp_tools_unsupported
+        && is_tool_unsupported_error(&error_message)
+    {
+        return handle_mcp_unsupported_error(app, ctx);
+    }
+
+    let input_area_height = app.input_area_height(ctx.term_width);
+    {
+        let mut conversation = app.conversation();
+        conversation.remove_trailing_empty_assistant_messages();
+        conversation.clear_pending_tool_calls();
+        conversation.add_app_message(AppMessageKind::Error, error_message);
+        let available_height =
+            conversation.calculate_available_height(ctx.term_height, input_area_height);
+        conversation.update_scroll_position(available_height, ctx.term_width);
+    }
+    app.end_streaming();
+    None
+}
+
+fn handle_mcp_unsupported_error(app: &mut App, ctx: AppActionContext) -> Option<AppCommand> {
+    app.session.mcp_tools_unsupported = true;
+    app.session.mcp_tools_enabled = false;
+
+    let input_area_height = app.input_area_height(ctx.term_width);
+    {
+        let mut conversation = app.conversation();
+        conversation.remove_trailing_empty_assistant_messages();
+        conversation.clear_pending_tool_calls();
+        conversation.add_app_message(
+            AppMessageKind::Warning,
+            "MCP tool-calling is enabled, but the currently selected model does not support it. It will be disabled until you switch models."
+                .to_string(),
+        );
+        let available_height =
+            conversation.calculate_available_height(ctx.term_height, input_area_height);
+        conversation.update_scroll_position(available_height, ctx.term_width);
+    }
+    app.end_streaming();
+
+    let base_messages = app.session.last_stream_api_messages_base.clone()?;
+
+    if ctx.term_width == 0 || ctx.term_height == 0 {
+        return None;
+    }
+
+    app.ui.focus_transcript();
+    app.enable_auto_scroll();
+
+    let input_area_height = app.input_area_height(ctx.term_width);
+    let (cancel_token, stream_id) = {
+        let mut conversation = app.conversation();
+        let (cancel_token, stream_id) = conversation.start_new_stream();
+        conversation.add_assistant_placeholder();
+        let available_height =
+            conversation.calculate_available_height(ctx.term_height, input_area_height);
+        conversation.update_scroll_position(available_height, ctx.term_width);
+        (cancel_token, stream_id)
+    };
+
+    Some(AppCommand::SpawnStream(app.build_stream_params(
+        base_messages,
+        cancel_token,
+        stream_id,
+    )))
+}
+
+fn is_tool_unsupported_error(message: &str) -> bool {
+    let lower = message.to_ascii_lowercase();
+    let mentions_tools = lower.contains("tools")
+        || lower.contains("tool_calls")
+        || lower.contains("tool call")
+        || lower.contains("function_call")
+        || lower.contains("function calling");
+    if !mentions_tools {
+        return false;
+    }
+
+    let unsupported_signals = [
+        "not supported",
+        "unsupported",
+        "unknown field",
+        "unknown parameter",
+        "unrecognized",
+        "unexpected field",
+        "invalid parameter",
+        "extra fields",
+        "does not support",
+    ];
+
+    unsupported_signals
+        .iter()
+        .any(|signal| lower.contains(signal))
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::core::app::actions::{AppCommand, StreamingAction};
+    use crate::core::app::actions::AppCommand;
     use crate::core::message::{ROLE_APP_ERROR, ROLE_ASSISTANT, ROLE_USER};
     use crate::utils::test_utils::create_test_app;
 
@@ -23,31 +118,19 @@ mod tests {
     }
 
     #[test]
-    fn stream_errored_drops_empty_assistant_placeholder() {
+    fn stream_error_trims_message_and_keeps_user_history() {
         let mut app = create_test_app();
         let ctx = default_ctx();
 
-        let command = super::super::handle_streaming_action(
+        let command = super::super::stream_lifecycle::spawn_stream_for_message(
             &mut app,
-            StreamingAction::SubmitMessage {
-                message: "Hello there".into(),
-            },
+            "Hello there".into(),
             ctx,
         );
 
-        let stream_id = match command {
-            Some(AppCommand::SpawnStream(params)) => params.stream_id,
-            _ => panic!("expected stream"),
-        };
+        assert!(matches!(command, Some(AppCommand::SpawnStream(_))));
 
-        let result = super::super::handle_streaming_action(
-            &mut app,
-            StreamingAction::StreamErrored {
-                message: " network failure ".into(),
-                stream_id,
-            },
-            ctx,
-        );
+        let result = handle_stream_error(&mut app, " network failure ".into(), ctx);
         assert!(result.is_none());
         assert!(app
             .ui
