+++ src/core/app/actions/streaming.rs
@@ -10,16 +10,11 @@ use crate::core::app::session::{
 use crate::core::app::ui_state::ToolPromptRequest;
 use crate::core::chat_stream::StreamParams;
 use crate::core::config::data::McpToolPayloadRetention;
-use crate::core::mcp_sampling::{
-    build_sampling_messages, serialize_sampling_params, summarize_sampling_request,
-};
+use crate::core::mcp_sampling::{serialize_sampling_params, summarize_sampling_request};
 use crate::core::message::{AppMessageKind, Message, ROLE_ASSISTANT, ROLE_USER};
 use crate::mcp::permissions::ToolPermissionDecision;
 use crate::mcp::{MCP_INSTANT_RECALL_TOOL, MCP_SESSION_MEMORY_SERVER_ID};
-use rust_mcp_schema::schema_utils::ServerJsonrpcRequest;
-use rust_mcp_schema::{
-    ClientCapabilities, ClientSampling, ContentBlock, PromptMessage, Role, RpcError,
-};
+use rust_mcp_schema::{ContentBlock, PromptMessage, Role, RpcError};
 use serde_json::{Map, Value};
 use tracing::debug;
 
@@ -115,533 +110,7 @@ pub(super) fn spawn_stream_for_message(
     message: String,
     ctx: AppActionContext,
 ) -> Option<AppCommand> {
-    if should_defer_for_mcp(app) {
-        app.session.pending_mcp_message = Some(message);
-        set_status_for_mcp_wait(app, ctx);
-        return None;
-    }
-
-    let params = prepare_stream_params_for_message(app, message, ctx);
-    Some(AppCommand::SpawnStream(params))
-}
-
-pub(super) fn retry_last_message(app: &mut App, ctx: AppActionContext) -> Option<AppCommand> {
-    prepare_retry_stream(app, ctx)
-}
-
-pub(super) fn refine_last_message(
-    app: &mut App,
-    prompt: String,
-    ctx: AppActionContext,
-) -> Option<AppCommand> {
-    prepare_refine_stream(app, prompt, ctx)
-}
-
-fn append_response_chunk(app: &mut App, chunk: &str, ctx: AppActionContext) {
-    if chunk.is_empty() {
-        return;
-    }
-
-    let input_area_height = app.input_area_height(ctx.term_width);
-    let mut conversation = app.conversation();
-    let available_height =
-        conversation.calculate_available_height(ctx.term_height, input_area_height);
-    conversation.append_to_response(chunk, available_height, ctx.term_width);
-}
-
-fn append_stream_app_message(
-    app: &mut App,
-    kind: AppMessageKind,
-    message: String,
-    ctx: AppActionContext,
-) {
-    let trimmed = message.trim();
-    if trimmed.is_empty() {
-        return;
-    }
-
-    let input_area_height = app.input_area_height(ctx.term_width);
-    let mut conversation = app.conversation();
-    conversation.add_app_message(kind, trimmed.to_string());
-    let available_height =
-        conversation.calculate_available_height(ctx.term_height, input_area_height);
-    conversation.update_scroll_position(available_height, ctx.term_width);
-}
-
-fn handle_mcp_init_completed(app: &mut App, ctx: AppActionContext) -> Option<AppCommand> {
-    app.session.mcp_init_complete = true;
-    app.session.mcp_init_in_progress = false;
-    app.end_mcp_operation_if_active();
-    if let Some(message) = app.session.pending_mcp_message.take() {
-        spawn_stream_for_message(app, message, ctx)
-    } else {
-        None
-    }
-}
-
-fn handle_mcp_send_without_tools(app: &mut App, ctx: AppActionContext) -> Option<AppCommand> {
-    if let Some(message) = app.session.pending_mcp_message.take() {
-        app.end_mcp_operation_if_active();
-        Some(AppCommand::SpawnStream(prepare_stream_params_for_message(
-            app, message, ctx,
-        )))
-    } else {
-        None
-    }
-}
-
-fn should_defer_for_mcp(app: &App) -> bool {
-    if app.session.mcp_tools_unsupported
-        || app.session.mcp_init_complete
-        || !app.session.mcp_init_in_progress
-    {
-        return false;
-    }
-
-    app.mcp.servers().any(|server| server.config.is_enabled())
-}
-
-fn set_status_for_mcp_wait(app: &mut App, ctx: AppActionContext) {
-    let input_area_height = app.input_area_height(ctx.term_width);
-    app.begin_mcp_operation();
-    let mut conversation = app.conversation();
-    let available_height =
-        conversation.calculate_available_height(ctx.term_height, input_area_height);
-    conversation.update_scroll_position(available_height, ctx.term_width);
-}
-
-fn append_tool_call_delta(app: &mut App, delta: crate::core::chat_stream::ToolCallDelta) {
-    let entry = app
-        .session
-        .pending_tool_calls
-        .entry(delta.index)
-        .or_insert_with(|| PendingToolCall {
-            id: None,
-            name: None,
-            arguments: String::new(),
-        });
-
-    if delta.id.is_some() {
-        entry.id = delta.id;
-    }
-    if delta.name.is_some() {
-        entry.name = delta.name;
-    }
-    if let Some(arguments) = delta.arguments {
-        entry.arguments.push_str(&arguments);
-    }
-}
-
-fn handle_stream_error(
-    app: &mut App,
-    message: String,
-    ctx: AppActionContext,
-) -> Option<AppCommand> {
-    let error_message = message.trim().to_string();
-    if app.session.mcp_tools_enabled
-        && !app.session.mcp_tools_unsupported
-        && is_tool_unsupported_error(&error_message)
-    {
-        return handle_mcp_unsupported_error(app, ctx);
-    }
-
-    let input_area_height = app.input_area_height(ctx.term_width);
-    {
-        let mut conversation = app.conversation();
-        conversation.remove_trailing_empty_assistant_messages();
-        conversation.clear_pending_tool_calls();
-        conversation.add_app_message(AppMessageKind::Error, error_message);
-        let available_height =
-            conversation.calculate_available_height(ctx.term_height, input_area_height);
-        conversation.update_scroll_position(available_height, ctx.term_width);
-    }
-    app.end_streaming();
-    None
-}
-
-fn handle_mcp_unsupported_error(app: &mut App, ctx: AppActionContext) -> Option<AppCommand> {
-    app.session.mcp_tools_unsupported = true;
-    app.session.mcp_tools_enabled = false;
-
-    let input_area_height = app.input_area_height(ctx.term_width);
-    {
-        let mut conversation = app.conversation();
-        conversation.remove_trailing_empty_assistant_messages();
-        conversation.clear_pending_tool_calls();
-        conversation.add_app_message(
-            AppMessageKind::Warning,
-            "MCP tool-calling is enabled, but the currently selected model does not support it. It will be disabled until you switch models."
-                .to_string(),
-        );
-        let available_height =
-            conversation.calculate_available_height(ctx.term_height, input_area_height);
-        conversation.update_scroll_position(available_height, ctx.term_width);
-    }
-    app.end_streaming();
-
-    let base_messages = app.session.last_stream_api_messages_base.clone()?;
-
-    if ctx.term_width == 0 || ctx.term_height == 0 {
-        return None;
-    }
-
-    app.ui.focus_transcript();
-    app.enable_auto_scroll();
-
-    let input_area_height = app.input_area_height(ctx.term_width);
-    let (cancel_token, stream_id) = {
-        let mut conversation = app.conversation();
-        let (cancel_token, stream_id) = conversation.start_new_stream();
-        conversation.add_assistant_placeholder();
-        let available_height =
-            conversation.calculate_available_height(ctx.term_height, input_area_height);
-        conversation.update_scroll_position(available_height, ctx.term_width);
-        (cancel_token, stream_id)
-    };
-
-    Some(AppCommand::SpawnStream(app.build_stream_params(
-        base_messages,
-        cancel_token,
-        stream_id,
-    )))
-}
-
-fn is_tool_unsupported_error(message: &str) -> bool {
-    let lower = message.to_ascii_lowercase();
-    let mentions_tools = lower.contains("tools")
-        || lower.contains("tool_calls")
-        || lower.contains("tool call")
-        || lower.contains("function_call")
-        || lower.contains("function calling");
-    if !mentions_tools {
-        return false;
-    }
-
-    let unsupported_signals = [
-        "not supported",
-        "unsupported",
-        "unknown field",
-        "unknown parameter",
-        "unrecognized",
-        "unexpected field",
-        "invalid parameter",
-        "extra fields",
-        "does not support",
-    ];
-
-    unsupported_signals
-        .iter()
-        .any(|signal| lower.contains(signal))
-}
-
-fn is_tool_error_payload(payload: &str) -> bool {
-    serde_json::from_str::<Value>(payload)
-        .ok()
-        .and_then(|value| value.get("isError").and_then(Value::as_bool))
-        .unwrap_or(false)
-}
-
-fn finalize_stream(app: &mut App, ctx: AppActionContext) -> Option<AppCommand> {
-    let input_area_height = app.input_area_height(ctx.term_width);
-    let pending_tool_calls = {
-        let mut conversation = app.conversation();
-        let pending = conversation.take_pending_tool_calls();
-        if !pending.is_empty() {
-            let available_height =
-                conversation.calculate_available_height(ctx.term_height, input_area_height);
-            conversation.update_scroll_position(available_height, ctx.term_width);
-        }
-        conversation.finalize_response();
-        pending
-    };
-
-    app.end_streaming();
-
-    if pending_tool_calls.is_empty() {
-        app.session.last_stream_api_messages = None;
-        return None;
-    }
-
-    prepare_tool_flow(app, pending_tool_calls, ctx)
-}
-
-fn handle_tool_permission_decision(
-    app: &mut App,
-    decision: ToolPermissionDecision,
-    ctx: AppActionContext,
-) -> Option<AppCommand> {
-    let prompt_tool = app.ui.tool_prompt().map(|prompt| prompt.tool_name.clone());
-    debug!(
-        decision = ?decision,
-        prompt_tool = prompt_tool.as_deref().unwrap_or("<none>"),
-        "Tool permission decision received"
-    );
-
-    if prompt_tool.as_deref() == Some(crate::mcp::MCP_SAMPLING_TOOL) {
-        let request = app.session.active_sampling_request.take()?;
-        return handle_sampling_permission_decision(app, request, decision, ctx);
-    }
-
-    if let Some(request) = app.session.active_tool_request.take() {
-        app.ui.cancel_tool_prompt();
-        app.clear_status();
-
-        match decision {
-            ToolPermissionDecision::AllowOnce => {}
-            ToolPermissionDecision::AllowSession | ToolPermissionDecision::Block => app
-                .mcp_permissions
-                .record(&request.server_id, &request.tool_name, decision),
-            ToolPermissionDecision::DenyOnce => {}
-        }
-
-        if matches!(
-            decision,
-            ToolPermissionDecision::DenyOnce | ToolPermissionDecision::Block
-        ) {
-            let message = match decision {
-                ToolPermissionDecision::Block => "Tool blocked by user.",
-                _ => "Tool denied by user.",
-            };
-            let server_label = resolve_server_label(app, &request.server_id);
-            let status = match decision {
-                ToolPermissionDecision::Block => ToolResultStatus::Blocked,
-                _ => ToolResultStatus::Denied,
-            };
-            let meta = tool_calls::ToolResultMeta::new(
-                Some(server_label),
-                Some(request.server_id.clone()),
-                request.tool_call_id.clone(),
-                Some(request.raw_arguments.clone()),
-            );
-            record_tool_result(
-                app,
-                &request.tool_name,
-                meta,
-                message.to_string(),
-                status,
-                ctx,
-            );
-            return advance_tool_queue(app, ctx);
-        }
-
-        app.session.active_tool_request = Some(request.clone());
-        if is_instant_recall_tool(&request.tool_name) {
-            return handle_instant_recall_tool_request(app, request, ctx);
-        }
-        set_status_for_tool_run(app, &request, ctx);
-        return Some(AppCommand::RunMcpTool(request));
-    }
-
-    let request = app.session.active_sampling_request.take()?;
-    handle_sampling_permission_decision(app, request, decision, ctx)
-}
-
-fn handle_mcp_server_request(
-    app: &mut App,
-    request: crate::mcp::events::McpServerRequest,
-    ctx: AppActionContext,
-) -> Option<AppCommand> {
-    debug!(
-        server_id = %request.server_id,
-        request_id = ?request.request.request_id(),
-        method = %request.request.method(),
-        "Received MCP server request"
-    );
-    if app.session.mcp_disabled {
-        debug!(
-            server_id = %request.server_id,
-            request_id = ?request.request.request_id(),
-            "MCP server request ignored (MCP disabled)"
-        );
-        return Some(AppCommand::SendMcpServerError {
-            server_id: request.server_id,
-            request_id: request.request.request_id().clone(),
-            error: RpcError::internal_error().with_message("MCP is disabled."),
-        });
-    }
-
-    let capabilities = ClientCapabilities {
-        sampling: Some(ClientSampling::default()),
-        ..ClientCapabilities::default()
-    };
-    if let Err(err) = capabilities.can_handle_request(&request.request) {
-        debug!(
-            server_id = %request.server_id,
-            request_id = ?request.request.request_id(),
-            error_code = err.code,
-            "MCP server request rejected (unsupported)"
-        );
-        return Some(AppCommand::SendMcpServerError {
-            server_id: request.server_id,
-            request_id: request.request.request_id().clone(),
-            error: err,
-        });
-    }
-
-    let request_id = request.request.request_id().clone();
-    let ServerJsonrpcRequest::CreateMessageRequest(create_request) = request.request else {
-        debug!(
-            server_id = %request.server_id,
-            request_id = ?request_id,
-            "MCP server request rejected (unexpected method)"
-        );
-        return Some(AppCommand::SendMcpServerError {
-            server_id: request.server_id,
-            request_id,
-            error: RpcError::method_not_found()
-                .with_message("Unsupported MCP request from server."),
-        });
-    };
-
-    let messages = match build_sampling_messages(&create_request) {
-        Ok(messages) => messages,
-        Err(err) => {
-            debug!(
-                server_id = %request.server_id,
-                request_id = ?create_request.id,
-                error = %err,
-                "MCP sampling request rejected (invalid params)"
-            );
-            return Some(AppCommand::SendMcpServerError {
-                server_id: request.server_id,
-                request_id: create_request.id.clone(),
-                error: RpcError::invalid_params().with_message(&err),
-            });
-        }
-    };
-
-    app.session
-        .pending_sampling_queue
-        .push_back(McpSamplingRequest {
-            server_id: request.server_id,
-            request: create_request,
-            messages,
-        });
-
-    debug!(
-        pending = app.session.pending_sampling_queue.len(),
-        active = app.session.active_sampling_request.is_some(),
-        "Queued MCP sampling request"
-    );
-
-    advance_sampling_queue(app, ctx)
-}
-
-fn handle_sampling_permission_decision(
-    app: &mut App,
-    request: McpSamplingRequest,
-    decision: ToolPermissionDecision,
-    ctx: AppActionContext,
-) -> Option<AppCommand> {
-    debug!(
-        server_id = %request.server_id,
-        request_id = ?request.request.id,
-        decision = ?decision,
-        "Sampling permission decision"
-    );
-    app.ui.cancel_tool_prompt();
-    app.clear_status();
-
-    match decision {
-        ToolPermissionDecision::AllowOnce => {}
-        ToolPermissionDecision::AllowSession | ToolPermissionDecision::Block => app
-            .mcp_permissions
-            .record(&request.server_id, crate::mcp::MCP_SAMPLING_TOOL, decision),
-        ToolPermissionDecision::DenyOnce => {}
-    }
-
-    if matches!(
-        decision,
-        ToolPermissionDecision::DenyOnce | ToolPermissionDecision::Block
-    ) {
-        let message = match decision {
-            ToolPermissionDecision::Block => "Sampling blocked by user.",
-            _ => "Sampling denied by user.",
-        };
-        return Some(AppCommand::SendMcpServerError {
-            server_id: request.server_id,
-            request_id: request.request.id.clone(),
-            error: RpcError {
-                code: -1,
-                message: message.to_string(),
-                data: None,
-            },
-        });
-    }
-
-    app.session.active_sampling_request = Some(request.clone());
-    set_status_for_sampling_run(app, &request, ctx);
-    Some(AppCommand::RunMcpSampling(Box::new(request)))
-}
-
-fn handle_mcp_sampling_finished(app: &mut App, ctx: AppActionContext) -> Option<AppCommand> {
-    app.session.active_sampling_request = None;
-    if let Some(request) = app.session.active_tool_request.clone() {
-        set_status_for_tool_run(app, &request, ctx);
-    } else {
-        app.end_mcp_operation_if_active();
-    }
-    advance_sampling_queue(app, ctx)
-}
-
-fn handle_tool_call_completed(
-    app: &mut App,
-    tool_name: String,
-    tool_call_id: Option<String>,
-    result: Result<String, String>,
-    ctx: AppActionContext,
-) -> Option<AppCommand> {
-    let request = app.session.active_tool_request.take();
-    let server_label = request
-        .as_ref()
-        .map(|req| resolve_server_label(app, &req.server_id));
-    app.end_mcp_operation_if_active();
-
-    match result {
-        Ok(payload) => {
-            let is_tool_error = is_tool_error_payload(&payload);
-            let mut meta = tool_calls::ToolResultMeta::new(
-                server_label,
-                request.as_ref().map(|req| req.server_id.clone()),
-                tool_call_id.clone(),
-                request.as_ref().map(|req| req.raw_arguments.clone()),
-            );
-            if is_tool_error {
-                meta.failure_kind = Some(ToolFailureKind::ToolError);
-            }
-            record_tool_result(
-                app,
-                &tool_name,
-                meta,
-                payload,
-                if is_tool_error {
-                    ToolResultStatus::Error
-                } else {
-                    ToolResultStatus::Success
-                },
-                ctx,
-            );
-        }
-        Err(err) => {
-            let mut meta = tool_calls::ToolResultMeta::new(
-                server_label,
-                request.as_ref().map(|req| req.server_id.clone()),
-                tool_call_id,
-                request.as_ref().map(|req| req.raw_arguments.clone()),
-            );
-            meta.failure_kind = Some(ToolFailureKind::ToolCallFailure);
-            record_tool_result(
-                app,
-                &tool_name,
-                meta,
-                format!("Tool call failure: {err}"),
-                ToolResultStatus::Error,
-                ctx,
-            );
-        }
-    }
-
-    advance_tool_queue(app, ctx)
+    stream_lifecycle::spawn_stream_for_message(app, message, ctx)
 }
 
 fn handle_mcp_prompt_completed(
